avoid crashes if test data not available
git-svn-id: https://zenmagick.svn.sourceforge.net/svnroot/zenmagick/trunk@1634 2e913b60-282b-11dc-bd19-815372732833

+++ zenmagick/plugins/request/zm_tests/tests/service/TestZMOrders.php
@@ -14,16 +14,20 @@ class TestZMOrders extends ZMTestCase {
      */
     public function testUpdateOrderStatus() {
         $order = ZMOrders::instance()->getOrderForId(1);
-        $order->setOrderStatusId(4);
-        ZMOrders::instance()->updateOrder($order);
-        $order = ZMOrders::instance()->getOrderForId(1);
-        $this->assertEqual(4, $order->getOrderStatusId());
-        $this->assertEqual('Update', $order->getStatusName());
-        $order->setOrderStatusId(2);
-        ZMOrders::instance()->updateOrder($order);
-        $order = ZMOrders::instance()->getOrderForId(1);
-        $this->assertEqual(2, $order->getOrderStatusId());
-        $this->assertEqual('Processing', $order->getStatusName());
+        if (null != $order) {
+            $order->setOrderStatusId(4);
+            ZMOrders::instance()->updateOrder($order);
+            $order = ZMOrders::instance()->getOrderForId(1);
+            $this->assertEqual(4, $order->getOrderStatusId());
+            $this->assertEqual('Update', $order->getStatusName());
+            $order->setOrderStatusId(2);
+            ZMOrders::instance()->updateOrder($order);
+            $order = ZMOrders::instance()->getOrderForId(1);
+            $this->assertEqual(2, $order->getOrderStatusId());
+            $this->assertEqual('Processing', $order->getStatusName());
+        } else {
+            $this->fail('test order not found');
+        }
     }
 
     /**
@@ -40,10 +44,14 @@ class TestZMOrders extends ZMTestCase {
      */
     public function testGetAccount() {
         $order = ZMOrders::instance()->getOrderForId(1);
-        $account = $order->getAccount();
-        $this->assertNotNull($account);
-        $this->assertNotNull($account->getLastName());
-        $this->assertNotNull($account->getEmail());
+        if (null != $order) {
+            $account = $order->getAccount();
+            $this->assertNotNull($account);
+            $this->assertNotNull($account->getLastName());
+            $this->assertNotNull($account->getEmail());
+        } else {
+            $this->fail('test order not found');
+        }
     }
 
     /**
@@ -53,21 +61,26 @@ class TestZMOrders extends ZMTestCase {
         $order = ZMOrders::instance()->getOrderForId(1);
         $this->assertNotNull($order);
 
-        $address = ZMLoader::make('Address');
-        $address->setFirstName('foo');
-        $address->setLastName('bar');
-        $address->setCompanyName('dooh inc.');
-        $address->setAddress('street 1');
-        $address->setSuburb('sub');
-        $address->setPostcode('12345');
-        $address->setCity('Christchurch');
-        $address->setState('Canterbury');
-        $address->setCountryId(153);
-        //address format is derived from country
+        if (null != $order) {
+            $address = ZMLoader::make('Address');
+            $address->setFirstName('foo');
+            $address->setLastName('bar');
+            $address->setCompanyName('dooh inc.');
+            $address->setAddress('street 1');
+            $address->setSuburb('sub');
+            $address->setPostcode('12345');
+            $address->setCity('Christchurch');
+            $address->setState('Canterbury');
+            $address->setCountryId(153);
+            //address format is derived from country
 
-        $order->setBillingAddress($address);
+            $order->setBillingAddress($address);
 
-        $this->assertEqual('12345', $order->get('billing_postcode'));
+            $this->assertEqual('12345', $order->get('billing_postcode'));
+        } else {
+            $this->fail('test order not found');
+        }
+
     }
 
     /**
@@ -89,18 +102,22 @@ class TestZMOrders extends ZMTestCase {
         $order = ZMOrders::instance()->getOrderForId(1);
         $this->assertNotNull($order);
 
-        $orderStatusHistory = $order->getOrderStatusHistory();
-        $this->assertNotNull($orderStatusHistory);
-        $this->assertTrue(is_array($orderStatusHistory));
-        $this->assertEqual(1, count($orderStatusHistory));
+        if (null != $order) {
+            $orderStatusHistory = $order->getOrderStatusHistory();
+            $this->assertNotNull($orderStatusHistory);
+            $this->assertTrue(is_array($orderStatusHistory));
+            $this->assertEqual(1, count($orderStatusHistory));
 
-        // check first entry
-        $orderStatus = $orderStatusHistory[0];
-        $this->assertEqual(1, $orderStatus->getId());
-        $this->assertEqual(1, $orderStatus->getOrderId());
-        $this->assertEqual('Pending', $orderStatus->getName());
-        $this->assertEqual(true, $orderStatus->isCustomerNotified());
-        $this->assertEqual(null, $orderStatus->getComment());
+            // check first entry
+            $orderStatus = $orderStatusHistory[0];
+            $this->assertEqual(1, $orderStatus->getId());
+            $this->assertEqual(1, $orderStatus->getOrderId());
+            $this->assertEqual('Pending', $orderStatus->getName());
+            $this->assertEqual(true, $orderStatus->isCustomerNotified());
+            $this->assertEqual(null, $orderStatus->getComment());
+        } else {
+            $this->fail('test order not found');
+        }
     }
 
     /**
@@ -110,31 +127,35 @@ class TestZMOrders extends ZMTestCase {
         $order = ZMOrders::instance()->getOrderForId(1);
         $this->assertNotNull($order);
 
-        $orderStatusHistory = $order->getOrderStatusHistory();
-        $this->assertNotNull($orderStatusHistory);
-        $this->assertTrue(is_array($orderStatusHistory));
-        $this->assertEqual(1, count($orderStatusHistory));
+        if (null != $order) {
+            $orderStatusHistory = $order->getOrderStatusHistory();
+            $this->assertNotNull($orderStatusHistory);
+            $this->assertTrue(is_array($orderStatusHistory));
+            $this->assertEqual(1, count($orderStatusHistory));
 
-        $newOrderStatus = ZMLoader::make('OrderStatus');
-        $newOrderStatus->setOrderId(1);
-        $newOrderStatus->setOrderStatusId(2);
-        $newOrderStatus = ZMOrders::instance()->createOrderStatusHistory($newOrderStatus);
-        // check for new primary key
-        $this->assertTrue(0 != $newOrderStatus->getId());
+            $newOrderStatus = ZMLoader::make('OrderStatus');
+            $newOrderStatus->setOrderId(1);
+            $newOrderStatus->setOrderStatusId(2);
+            $newOrderStatus = ZMOrders::instance()->createOrderStatusHistory($newOrderStatus);
+            // check for new primary key
+            $this->assertTrue(0 != $newOrderStatus->getId());
 
-        $orderStatusHistory = $order->getOrderStatusHistory();
-        $this->assertNotNull($orderStatusHistory);
-        $this->assertTrue(is_array($orderStatusHistory));
-        $this->assertEqual(2, count($orderStatusHistory));
-        // check created entry
-        $createdOrderStatus = $orderStatusHistory[1];
-        // make sure this is set
-        $this->assertEqual('Processing', $createdOrderStatus->getName());
-        $this->assertNotNull($createdOrderStatus->getDateAdded());
+            $orderStatusHistory = $order->getOrderStatusHistory();
+            $this->assertNotNull($orderStatusHistory);
+            $this->assertTrue(is_array($orderStatusHistory));
+            $this->assertEqual(2, count($orderStatusHistory));
+            // check created entry
+            $createdOrderStatus = $orderStatusHistory[1];
+            // make sure this is set
+            $this->assertEqual('Processing', $createdOrderStatus->getName());
+            $this->assertNotNull($createdOrderStatus->getDateAdded());
 
-        // clean up
-        $sql = "DELETE FROM ".TABLE_ORDERS_STATUS_HISTORY." WHERE orders_status_history_id = :orderStatusHistoryId";
-        ZMRuntime::getDatabase()->update($sql, array('orderStatusHistoryId' => $newOrderStatus->getId()), TABLE_ORDERS_STATUS_HISTORY);
+            // clean up
+            $sql = "DELETE FROM ".TABLE_ORDERS_STATUS_HISTORY." WHERE orders_status_history_id = :orderStatusHistoryId";
+            ZMRuntime::getDatabase()->update($sql, array('orderStatusHistoryId' => $newOrderStatus->getId()), TABLE_ORDERS_STATUS_HISTORY);
+        } else {
+            $this->fail('test order not found');
+        }
     }
 
     /**
@@ -144,17 +165,21 @@ class TestZMOrders extends ZMTestCase {
         $order = ZMOrders::instance()->getOrderForId(1);
         $this->assertNotNull($order);
 
-        $totals = $order->getOrderTotals();
-        $this->assertNotNull($totals);
-        $this->assertTrue(is_array($totals));
-        $this->assertEqual(3, count($totals));
-        
-        // test total total
-        $total = $totals['ot_total'];
-        $this->assertEqual('Total:', $total->getName());
-        $this->assertEqual('$52.49', $total->getValue());
-        $this->assertEqual(52.49, $total->getAmount());
-        $this->assertEqual('ot_total', $total->getType());
+        if (null != $order) {
+            $totals = $order->getOrderTotals();
+            $this->assertNotNull($totals);
+            $this->assertTrue(is_array($totals));
+            $this->assertEqual(3, count($totals));
+            
+            // test total total
+            $total = $totals['ot_total'];
+            $this->assertEqual('Total:', $total->getName());
+            $this->assertEqual('$52.49', $total->getValue());
+            $this->assertEqual(52.49, $total->getAmount());
+            $this->assertEqual('ot_total', $total->getType());
+        } else {
+            $this->fail('test order not found');
+        }
     }
 
     /*
@@ -164,14 +189,18 @@ class TestZMOrders extends ZMTestCase {
         $order = ZMOrders::instance()->getOrderForId(109);
         $this->assertNotNull($order);
 
-        $items = $order->getOrderItems();
-        $this->assertEqual(1, count($items));
-        $item = $items[0];
-        $this->assertEqual(1, $item->getProductId());
-        $this->assertEqual(1, $item->getQty());
-        $this->assertNotNull($item->getTaxRate());
-        $this->assertEqual(10.00, $item->getTaxRate()->getRate());
-        $this->assertEqual(2, count($item->getAttributes()));
+        if (null != $order) {
+            $items = $order->getOrderItems();
+            $this->assertEqual(1, count($items));
+            $item = $items[0];
+            $this->assertEqual(1, $item->getProductId());
+            $this->assertEqual(1, $item->getQty());
+            $this->assertNotNull($item->getTaxRate());
+            $this->assertEqual(10.00, $item->getTaxRate()->getRate());
+            $this->assertEqual(2, count($item->getAttributes()));
+        } else {
+            $this->fail('test order not found');
+        }
     }
 
     /*
@@ -181,29 +210,33 @@ class TestZMOrders extends ZMTestCase {
         $order = ZMOrders::instance()->getOrderForId(109);
         $this->assertNotNull($order);
 
-        $items = $order->getOrderItems();
-        $this->assertEqual(1, count($items));
-        $item = $items[0];
-        $attributes = $item->getAttributes();
-        $this->assertEqual(2, count($attributes));
+        if (null != $order) {
+            $items = $order->getOrderItems();
+            $this->assertEqual(1, count($items));
+            $item = $items[0];
+            $attributes = $item->getAttributes();
+            $this->assertEqual(2, count($attributes));
 
-        if (2 == count($attributes)) {
-            // expect productId 1 with model:premium and memory:16MB
-            $attribute = $attributes[0];
-            $this->assertEqual('Model', $attribute->getName());
-            $values = $attribute->getValues();
-            $this->assertEqual(1, count($values));
-            $value = $values[0];
-            $this->assertEqual('Premium', $value->getName());
+            if (2 == count($attributes)) {
+                // expect productId 1 with model:premium and memory:16MB
+                $attribute = $attributes[0];
+                $this->assertEqual('Model', $attribute->getName());
+                $values = $attribute->getValues();
+                $this->assertEqual(1, count($values));
+                $value = $values[0];
+                $this->assertEqual('Premium', $value->getName());
 
-            $attribute = $attributes[1];
-            $this->assertEqual('Memory', $attribute->getName());
-            $values = $attribute->getValues();
-            $this->assertEqual(1, count($values));
-            $value = $values[0];
-            $this->assertEqual('16 mb', $value->getName());
+                $attribute = $attributes[1];
+                $this->assertEqual('Memory', $attribute->getName());
+                $values = $attribute->getValues();
+                $this->assertEqual(1, count($values));
+                $value = $values[0];
+                $this->assertEqual('16 mb', $value->getName());
+            } else {
+                $this->fail('missing attributes');
+            }
         } else {
-            $this->fail('missing attributes');
+            $this->fail('test order not found');
         }
     }
